Update to python-linstor-1.19.0
This adds source archive and spec file for python-linstor v1.19.0.

The spec file targets both xcp-ng 8.2 and 8.3, using python 2 and python
3 respectively. In order to build the package using python 3 for xcp-ng
8.3, the spec file relies on the presence of the python3 executable
which is only present in the xcp-ng 8.3 build container. The python 2
version for xcp-ng 8.2 is built otherwise.

The setup.py file is patched when building using python 2 for an
extras_require entry not recognized by the old setuptool version used in
the 8.2 build container.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>